centralized swap conversion rate calculation for consistency
